inventory_management/requirements.txt
Django==5.1.1
gunicorn==23.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dj-database-url==2.2.0
# Any other packages you use in your project
